typeerror bug fix

+++ lib/bill/delivered.dart
@@ -24,7 +24,7 @@ class DeliveredScreen extends StatefulWidget {
   final String meterno;
   final String lastpay;
   final double closingb;
-  final int lastpayamt;
+  final double lastpayamt;
   final String dropdownValue;
   final String geolat;
   final String geolong;
